feat(async) empty queues on nginx worker exit
### Summary

Previously the async library did not prevent jobs to be queued when
nginx worker was exiting. This commit prevents further jobs to be
added to queue when nginx worker is exiting.

It also changes code so that all the queued job are executed before
shutting down the threads or timers.